+++ src/main/java/com/t09g01/projeto/view/Viewer.java
@@ -23,7 +23,7 @@ public class Viewer {
 
 
 
-    public void draw(GUI gui, int a, int b){                           // a e b representam as coordenadas de início onde a imagem deve ser desenhada na tela
+    public void draw(GUI gui, double a, double b){                           // a e b representam as coordenadas de início onde a imagem deve ser desenhada na tela
         for (int y = 0; y < image.getHeight(); y++) {                  // x e y são as coordenadas dos pixels dentro da imagem
             for (int x = 0; x < image.getWidth(); x++) {
                 int rgb = image.getRGB(x, y);                         // valor do RGB do pixel
@@ -31,7 +31,7 @@ public class Viewer {
                     continue;
                 }
                 TextColor.RGB color = getLanternaColor(rgb);          // mapear para cor no Lanterna
-                gui.drawPixel(a + x,b + y, color);              // desenhar o pixel
+                gui.drawPixel((int)a + x,(int)b + y, color);              // desenhar o pixel
             }
         }
     }
